move to src / tests directory structure and lint with ruff
